split logical and physical graph classes
