+++ daliuge-translator/dlg/dropmake/pg_manager.py
@@ -27,8 +27,8 @@ import threading, json
 
 import numpy as np
 
-from .pg_generator import GraphException
-from .scheduler import DAGUtil, SchedulerException
+from dlg.dropmake.lg import GraphException
+from dlg.dropmake.scheduler import DAGUtil, SchedulerException
 
 
 MAX_PGT_FN_CNT = 300
