+++ daliuge-translator/test/dropmake/test_pg_gen.py
@@ -21,14 +21,9 @@
 
 import unittest, pkg_resources
 
-from dlg.dropmake.pg_generator import (
-    LG,
-    PGT,
-    MetisPGTP,
-    MySarkarPGTP,
-    MinNumPartsPGTP,
-    GPGTNoNeedMergeException,
-)
+from dlg.dropmake.lg import LG
+from dlg.dropmake.pgt import PGT, GPGTNoNeedMergeException
+from dlg.dropmake.pgtp import MetisPGTP, MySarkarPGTP, MinNumPartsPGTP
 from dlg.dropmake import pg_generator
 from dlg.translator.tool_commands import unroll
